fix(agent): extend app-server catalog bounds
